requirements.txt
streamlit==1.29.0
requests==2.31.0
plotly==5.18.0
pandas==2.1.4
pillow==10.1.0
numpy==1.26.2
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
tensorflow==2.15.0
scikit-learn==1.3.2
opencv-python-headless==4.8.1.78